fix(agent): flush dns cache during initialization (#120)
Since our agent is installing some apt packages, some of the apt sources
related domains/ips get cached on the runner. This cache prevents our
agent from correctly working if apt is used immediately after the agent
is installed, as it can be seen from the test-integration job in our CI
being
[flaky](https://github.com/bullfrogsec/bullfrog/actions/runs/10831682859/job/30054328993).

This PR is also fixing another issue with the test-integration with
virtualbox. Using virtualbox as the VM provider in vagrant started to
failed recently with the latest version of the runner image for
ubuntu-22.04
([ref](https://github.com/actions/runner-images/issues/10678)).

The two are fixed together since it is the only way to make the
test-integration CI job to pass reliably.